Use typescript where possible
